convert sync hotspots into node job
